Revert "Add Gemini CLI provider via shared ACP abstraction (#862)"
This reverts commit bcbd913b8e186ad1d2b1db50ac8344fd80debfe4.
