+++ src/fido/provider_factory.py
@@ -8,7 +8,6 @@ from pathlib import Path
 from fido.claude import ClaudeAPI, ClaudeClient, ClaudeCode
 from fido.config import RepoConfig
 from fido.copilotcli import CopilotCLI, CopilotCLIAPI, CopilotCLIClient
-from fido.gemini import Gemini, GeminiAPI, GeminiClient
 from fido.provider import (
     PromptSession,
     Provider,
@@ -36,8 +35,6 @@ class DefaultProviderFactory:
                     api = ClaudeAPI()
                 case ProviderID.COPILOT_CLI:
                     api = CopilotCLIAPI()
-                case ProviderID.GEMINI:
-                    api = GeminiAPI()
                 case _:
                     raise ValueError(f"unsupported provider: {repo_cfg.provider}")
             self._apis[repo_cfg.provider] = api
@@ -64,15 +61,6 @@ class DefaultProviderFactory:
             case ProviderID.COPILOT_CLI:
                 return CopilotCLI(
                     agent=CopilotCLIClient(
-                        session_system_file=self._session_system_file,
-                        work_dir=work_dir,
-                        repo_name=repo_name,
-                        session=session,
-                    )
-                )
-            case ProviderID.GEMINI:
-                return Gemini(
-                    agent=GeminiClient(
                         session_system_file=self._session_system_file,
                         work_dir=work_dir,
                         repo_name=repo_name,
