+++ tests/test_copilot_hold_for_handler.py
@@ -8,12 +8,10 @@ import pytest
 
 from fido import provider
 from fido.copilotcli import CopilotCLIClient, CopilotCLISession
-from fido.provider import ProviderID
 
 
 class FakeRuntime:
     def __init__(self) -> None:
-        self.provider_id = ProviderID.COPILOT_CLI
         self._session_id = "sess-created"
         self.cancel_calls: list[str] = []
         self.dropped_session_count = 0
